add method and use case for deep cloning an assessment
I'm using the model of 'ownership' for determining how assessments and their items can be edited. Users can edit their own Assessments and AssessmentItems, but to make changes, the user must clone the Assessment and it's AssessmentItems (deep cloning) and then make changes to it.

This is analogous to cloning a repo in Git, where we make a copy tht has a link back to the repo from which it was cloned, in order to be able to pull in upstream changes if desired.